DNS_lookup/requirements.txt
dns.resolver
